- correctly get local NTP server info - fix some bugs in the socket code
svn path=/trunk/; revision=26598

+++ reactos/dll/cpl/timedate/monthcal.c
@@ -176,7 +176,7 @@ MonthCalUpdate(IN PMONTHCALWND infoPtr)
     pDayEnd = pDay + MonthLength;
     while (pDay != pDayEnd)
     {
-        *(pDay++) = ++d;
+        *(pDay++) = (BYTE)++d;
     }
 
     /* repaint the control */
